# Teaching-HEIGVD-AMT-2016-Gamification/spring-server/src/main/resources/api-spec.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: Gamification Platform API
  description: |
   **This is the documentation of the AMT Gamification Platform API**
host: localhost:8080
basePath: /api
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /pointScales:
    post:
      parameters:
        - name: pointScaleDTO
          in: body
          description: The info required to create a pointScale
          schema:
            $ref: '#/definitions/pointScalePost'
          required: true
        - name: token
          in: header
          type: string
          description: token
      responses:
        201:
          description: Make a new pointScale
  
    get:
      responses:
        200:
          description:  PointScale list
          schema:
            title: pointScaleDTO
            type: array
            items:
              $ref: '#/definitions/pointScaleGet'
  /pointScales/{id}:
    get:
      parameters:
        - name: id
          in: path
          type: integer
          format: int64
          description: id of the pointScale
          required: true
      responses:
        200:
          description: Info of the poinstScale
          schema:
            title: pointScaleDTO
            type: object
            items:
              $ref: '#/definitions/pointScaleGet'
        404:
          description: PointScale not found
          
    put:
      parameters:
        - name: id
          in: path
          type: integer
          format: int64
          description: id of the pointScale to modify
          required: true
        - name: token
          in: header
          type: string
          description: token
        - name: pointScaleDTO
          in: body
          description: The new values for an existing pointScale
          schema:
            $ref: '#/definitions/pointScalePost'
          required: true
      responses:
        200:
          description: PointScale modified successfully
          schema:
            title: pointScaleDTO
            type: object
            items:
              $ref: '#/definitions/pointScaleGet'
        404:
          description: PointScale not found
          
    delete:
      parameters:
        - name: id
          in: path
          type: integer
          format: int64
          description: id of the pointScale to delete
          required: true
      responses:
        200:
          description: PointScale deleted successfully
        404:
          description: PointScale not found
          
#-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
# rule
#-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-            
  /rules:
    post:
      parameters:
        - name: ruleDTO
          in: body
          description: The info required to create a rule
          schema:
            $ref: '#/definitions/rulePost'
          required: true
        - name: token
          in: header
          type: string
          description: token
      responses:
        201:
          description: Rule created successfully
  
    get:
      responses:
        200:
          description:  List all rules
          schema:
            title: ruleDTO
            type: array
            items:
              $ref: '#/definitions/ruleGet'
  /rules/{id}:
    get:
      parameters:
        - name: id
          in: path
          type: integer
          format: int64
          description: id of the rule
          required: true
      responses:
        200:
          description: Rule infos
          schema:
            title: ruleDTO
            type: object
            items:
              $ref: '#/definitions/ruleGet'
        404:
          description: Rule not found
          
    put:
      parameters:
        - name: id
          in: path
          type: integer
          format: int64
          description: id of the rule to modify
          required: true
        - name: ruleDTO
          in: body
          description: The new values for an existing rule
          schema:
            $ref: '#/definitions/rulePost'
          required: true
      responses:
        200:
          description: Rule modified successfully
          schema:
            title: ruleDTO
            type: object
            items:
              $ref: '#/definitions/ruleGet'
        404:
          description: Rule not found
          
    delete:
      parameters:
        - name: id
          in: path
          type: integer
          format: int64
          description: id of the rule to delete
          required: true
      responses:
        200:
          description: Rule deleted successfully
        404:
          description: Rule not found
definitions:
...
  ruleGet:
    type: object
    properties:
      applicationId:
        type: integer
        format: int64
      applicationName:
        type: string
      eventType:
        type: string
      awardBadgeId:
        type: integer
        format: int64
      awardPointScaleId:
        type: integer
        format: int64
      numberOfPoint:
        type: integer
        format: int64
  rulePost:
    type: object
    properties:
      conditions:
        type: array
        items:
          type: string
      eventType:
        type:  string
      awardBadgeId:
        type: integer
        format: int64
      awardPointScaleId:
        type: integer
        format: int64
      point:
        type: integer
        format: int64
      properties:
        type: array
        items:
          type: integer
          format: int64
...
  pointScalePost:
    type: object
    properties:
      name:
        type: string
  pointScaleGet:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      applicationId:
        type: integer
        format: int64
      apllicationName:
        type: string
      
      
    
#-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
# event
#-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-         
...